Rewrite `mezzaluna-feature-loader`
Reorganize the crate for maintainability and ease of Cargo feature
composition. The crate is now fully documented with many more examples.

The `loader` module is moved to `loaders`, to mirror the plural used in
e.g. `std::collections`. No loaders are exported at the crate root; they
are only available by pathing into the `loaders` module.

`Feature` and the cfg'ing associated with them are relocated to the
`crate::feature` module. `crate::feature::Feature` is its only public
type; it is exported at the crate root.

The following APIs and types have been removed:

- `mezzaluna_feature_loader::Loader`
- `LoadedFeatures::insert_in_memory_feature` (replaced by exposing the
  underlying `Feature` struct)
- `LoadedFeatures::insert_on_disk_feature` (replaced by exposing the
  underlying `Feature` struct)
- `LoadedFeatures::clear` (this type documents it is append-only, which
  clear violates)

The API additions and breaking changes have been bundled into version
0.4.0.
