+++ mezzaluna-feature-loader/src/feature/memory.rs
@@ -1,0 +1,16 @@
+use std::path::{Path, PathBuf};
+
+#[derive(Debug, Hash, PartialEq, Eq)]
+pub struct Feature {
+    path: PathBuf,
+}
+
+impl Feature {
+    pub fn with_path(path: PathBuf) -> Self {
+        Self { path }
+    }
+
+    pub fn path(&self) -> &Path {
+        &*self.path
+    }
+}
